Decouple code for server startup, lifecycle hooks, FastAPI app, and routes
