+++ shortfin/python/shortfin_apps/llm/lifecycle_hooks.py
@@ -1,0 +1,37 @@
+# Copyright 2024 Advanced Micro Devices, Inc.
+#
+# Licensed under the Apache License v2.0 with LLVM Exceptions.
+# See https://llvm.org/LICENSE.txt for license information.
+# SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+
+from contextlib import asynccontextmanager
+import logging
+from typing import Any
+from fastapi import FastAPI
+
+from .components.manager import SystemManager
+
+sysman: SystemManager
+services: dict[str, Any] = {}
+
+
+@asynccontextmanager
+async def lifespan(app: FastAPI):
+    global sysman
+    global services
+
+    sysman.start()
+    try:
+        for service_name, service in services.items():
+            logging.info("Initializing service '%s': %r", service_name, service)
+            service.start()
+    except:
+        sysman.shutdown()
+        raise
+    yield
+    try:
+        for service_name, service in services.items():
+            logging.info("Shutting down service '%s'", service_name)
+            service.shutdown()
+    finally:
+        sysman.shutdown()
