+++ shortfin/python/shortfin_apps/llm/routes/__init__.py
@@ -1,0 +1,4 @@
+from .application import application_router
+from .generate import generation_router
+
+__all__ = ["application_router", "generation_router"]
