+++ shortfin/python/shortfin_apps/llm/routes/application.py
@@ -1,0 +1,8 @@
+from fastapi import APIRouter, Response
+
+application_router = APIRouter()
+
+
+@application_router.get("/health")
+async def health() -> Response:
+    return Response(status_code=200)
