+++ shortfin/python/shortfin_apps/llm/routes/generate.py
@@ -1,0 +1,19 @@
+from fastapi import APIRouter, Request
+
+from shortfin.interop.fastapi import FastAPIResponder
+
+from ..components.generate import ClientGenerateBatchProcess
+from ..components.io_struct import GenerateReqInput
+from ..lifecycle_hooks import services
+
+generation_router = APIRouter()
+
+
+@generation_router.post("/generate")
+@generation_router.put("/generate")
+async def generate_request(gen_req: GenerateReqInput, request: Request):
+    service = services["default"]
+    gen_req.post_init()
+    responder = FastAPIResponder(request)
+    ClientGenerateBatchProcess(service, gen_req, responder).launch()
+    return await responder.response
