+++ shortfin/python/shortfin_apps/llm/server.py
@@ -4,8 +4,6 @@
 # See https://llvm.org/LICENSE.txt for license information.
 # SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-from typing import Any
-
 import argparse
 import logging
 from pathlib import Path
@@ -15,17 +13,11 @@ import uvicorn.logging
 
 # Import first as it does dep checking and reporting.
 from shortfin import ProgramIsolation
-from shortfin.interop.fastapi import FastAPIResponder
-
-from contextlib import asynccontextmanager
-
-from fastapi import FastAPI, Request, Response
 import uvicorn
 
-
-from .components.generate import ClientGenerateBatchProcess
+from . import lifecycle_hooks
+from .application import get_app
 from .components.config_struct import ModelParams, ServerParams
-from .components.io_struct import GenerateReqInput
 from .components.manager import SystemManager
 from .components.service import GenerateService
 from .components.tokenizer import Tokenizer
@@ -59,47 +51,6 @@ UVICORN_LOG_CONFIG = {
         },
     },
 }
-
-
-@asynccontextmanager
-async def lifespan(app: FastAPI):
-    sysman.start()
-    try:
-        for service_name, service in services.items():
-            logging.info("Initializing service '%s': %r", service_name, service)
-            service.start()
-    except:
-        sysman.shutdown()
-        raise
-    yield
-    try:
-        for service_name, service in services.items():
-            logging.info("Shutting down service '%s'", service_name)
-            service.shutdown()
-    finally:
-        sysman.shutdown()
-
-
-sysman: SystemManager
-services: dict[str, Any] = {}
-app = FastAPI(lifespan=lifespan)
-
-
-@app.get("/health")
-async def health() -> Response:
-    return Response(status_code=200)
-
-
-async def generate_request(gen_req: GenerateReqInput, request: Request):
-    service = services["default"]
-    gen_req.post_init()
-    responder = FastAPIResponder(request)
-    ClientGenerateBatchProcess(service, gen_req, responder).launch()
-    return await responder.response
-
-
-app.post("/generate")(generate_request)
-app.put("/generate")(generate_request)
 
 
 def get_eos_from_tokenizer_config(json_path):
@@ -142,13 +93,13 @@ def configure(args) -> SystemManager:
     )
     sm.load_inference_module(args.vmfb)
     sm.load_inference_parameters(*args.parameters, parameter_scope="model")
-    services[sm.name] = sm
+    lifecycle_hooks.services[sm.name] = sm
     return sysman
 
 
 def main(argv, log_config=uvicorn.config.LOGGING_CONFIG):
     parser = argparse.ArgumentParser()
-    parser.add_argument("--host", type=str, default=None)
+    parser.add_argument("--host", type=str, default="0.0.0.0")
     parser.add_argument("--port", type=int, default=8000)
     parser.add_argument(
         "--root-path",
@@ -243,11 +194,10 @@ def main(argv, log_config=uvicorn.config.LOGGING_CONFIG):
             args.tokenizer_json.stem + "_config.json"
         )
         args.tokenizer_config_json = inferred_tokenizer_config_path
-    global sysman
-    sysman = configure(args)
+    lifecycle_hooks.sysman = configure(args)
 
     uvicorn.run(
-        app,
+        get_app(),
         host=args.host,
         port=args.port,
         log_config=log_config,
